requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1
django-filter==2.4.0
django-heroku==0.3.1
django-storages==1.10.1
dropbox==10.8.0
gunicorn==20.0.4
idna==2.10
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.11
whitenoise==5.2.0
